Refactor codebase to drop so much `#ifdef _WIN32` (#2142)
* refactor: abstract path utilities to use virtual interfaces

Replace scattered #ifdef blocks in path utilities with C++ inheritance
and virtual functions, following the TrayInterface pattern.

This establishes the pattern for remaining phases (process management,
system metrics, archive extraction).

Replace platform-specific #ifdef blocks for archive extraction with C++
inheritance pattern.

Extract platform-specific process spawning, lifecycle, and utilities from
process_manager.cpp into dedicated platform implementations following the
established TrayInterface pattern.

Extract all Windows process management code from process_manager.cpp inline
implementation into dedicated process_windows.cpp platform file, completing
the platform abstraction pattern across all three platforms.

Changes:
- Create PathPlatform abstract interface with platform-specific methods
- Implement WindowsPathPlatform, LinuxPathPlatform, MacOSPathPlatform
- Refactor path_utils.cpp to delegate to platform implementation
- Update CMakeLists.txt for platform-specific source selection
- Create SystemMetricsPlatform abstract interface
- Implement LinuxMetricsPlatform, WindowsMetricsPlatform, MacOSMetricsPlatform
- Refactor server.cpp to delegate metrics collection to platform
- Add metrics_platform_ member to Server class
- Update CMakeLists.txt for platform-specific metrics sources
- Create ArchivePlatform abstract interface
- Implement WindowsArchivePlatform and UnixArchivePlatform
- Refactor BackendUtils::extract_zip() and extract_tarball() to delegate
- Windows: Native tar.exe with PowerShell Expand-Archive fallback for ZIP
- Unix: unzip for ZIP, tar for tarballs (.tar.gz, .tar.xz, etc.)
- Update CMakeLists.txt for platform-specific archive sources
- process_manager.cpp: Delegate to platform on Unix/macOS for start_process(),
  stop_process(), is_running(), find_free_port()
- Windows code remains inline (future extraction possible)
- CLI CMakeLists.txt: Add process platform sources to link correctly
- process_manager.cpp: All functions now delegate to platform on all OS
  (start_process, stop_process, is_running, get_exit_code, wait_for_exit,
   run_with_output, find_free_port, run_command, kill_process)
- Windows helper functions (escape_windows_arg, output_filter_thread,
  build_windows_environment_block) moved to process_windows.cpp
- Old inline implementations wrapped in #if 0 blocks for reference
- CMakeLists.txt: Add process_windows.cpp to Windows builds
- CLI CMakeLists.txt: Add process_windows.cpp to CLI Windows builds

Results:
- Reduced #ifdef blocks in path_utils.cpp from 15 to 5 (67% reduction)
- Remaining #ifdef only for includes and small utility functions
- Public API unchanged (backward compatible)
- Build verified successful on Linux
- Reduced #ifdef blocks in server.cpp (system metrics functions)
- CPU usage: Delegates to platform (Linux /proc/stat, Windows GetSystemTimes)
- Memory: Delegates to platform (Linux /proc/meminfo, Windows GlobalMemoryStatusEx, macOS sysctl)
- GPU/VRAM: Linux AMD sysfs, unsupported on Windows/macOS
- NPU: Linux DRM IOCTL sensor API, unsupported on Windows/macOS
- Remaining #ifdef in server.cpp only for includes and small utilities
- Build verified successful on Linux
- CLI updated to include platform files
- Reduced #ifdef blocks in backend_utils.cpp from 14 to 9 (36% reduction)
- Remaining #ifdef only for includes, .7z support, and TheRock paths
- extract_seven_zip() unchanged (Windows-specific .7z using bsdtar)
- Build verified successful on Linux

* Fix for code review feedback

* fix: restore Windows header discipline and HF_HOME path handling

- process_windows.cpp: Add WIN32_LEAN_AND_MEAN, NOMINMAX, _WINSOCKAPI_
  guards, include winsock2.h before windows.h, and undef ERROR after
  Windows headers — matching the setup previously in process_manager.cpp
  before platform files were made self-contained. Prevents ERROR macro
  conflicts with LOG(ERROR, ...) and winsock2.h ordering problems.

- path_utils.cpp: Restore platform-specific string append for HF_HOME
  ('\hub' on Windows, '/hub' elsewhere) instead of fs::path division,
  preserving correct behavior for non-ASCII Windows paths.

* Avoid transitive includes
