+++ src/cpp/cli/CMakeLists.txt
@@ -98,6 +98,24 @@ set(COMMON_SOURCES
     pi_profile.cpp
 )
 
+# Add platform-specific sources
+if(WIN32)
+    list(APPEND COMMON_SOURCES
+        ../server/utils/platform/path_windows.cpp
+        ../server/utils/platform/process_windows.cpp
+    )
+elseif(APPLE)
+    list(APPEND COMMON_SOURCES
+        ../server/utils/platform/path_macos.cpp
+        ../server/utils/platform/process_macos.cpp
+    )
+elseif(UNIX)
+    list(APPEND COMMON_SOURCES
+        ../server/utils/platform/path_linux.cpp
+        ../server/utils/platform/process_unix.cpp
+    )
+endif()
+
 # ============================================================
 # lemonade - HTTP client CLI
 # ============================================================
