+++ src/cpp/server/server.cpp
@@ -198,7 +198,8 @@ static const json MIME_TYPES = {
 Server::Server(std::shared_ptr<RuntimeConfig> config, const std::string& cache_dir)
     : config_(config),
       cache_dir_(cache_dir),
-      port_(config->port()), running_(false), udp_beacon_() {
+      port_(config->port()), running_(false), udp_beacon_(),
+      metrics_platform_(create_metrics_platform()) {
 
     // Set global HttpClient timeout
     utils::HttpClient::set_default_timeout(config->global_timeout());
@@ -3805,295 +3806,30 @@ void Server::handle_system_info(const httplib::Request& req, httplib::Response& 
 
 // Get CPU usage percentage
 double Server::get_cpu_usage() {
-#ifdef __linux__
-    // Linux: Parse /proc/stat for system-wide CPU usage
-    std::lock_guard<std::mutex> lock(cpu_stats_mutex_);
-
-    std::ifstream stat_file("/proc/stat");
-    if (!stat_file.is_open()) {
-        return -1.0;
-    }
-
-    std::string line;
-    std::getline(stat_file, line);
-    stat_file.close();
-
-    // Parse: "cpu  user nice system idle iowait irq softirq steal"
-    std::istringstream iss(line);
-    std::string cpu_label;
-    uint64_t user, nice, system, idle, iowait, irq, softirq, steal;
-
-    iss >> cpu_label >> user >> nice >> system >> idle >> iowait >> irq >> softirq >> steal;
-
-    uint64_t total_idle = idle + iowait;
-    uint64_t total_active = user + nice + system + irq + softirq + steal;
-    uint64_t total = total_idle + total_active;
-
-    if (last_cpu_stats_.total > 0) {
-        uint64_t idle_diff = total_idle - last_cpu_stats_.total_idle;
-        uint64_t total_diff = total - last_cpu_stats_.total;
-
-        last_cpu_stats_.total_idle = total_idle;
-        last_cpu_stats_.total = total;
-
-        if (total_diff > 0) {
-            return ((total_diff - idle_diff) * 100.0) / total_diff;
-        }
-    }
-
-    last_cpu_stats_.total_idle = total_idle;
-    last_cpu_stats_.total = total;
-    return 0.0; // First call, no delta yet
-
-#elif defined(_WIN32)
-    // Windows: Use GetSystemTimes for system-wide CPU usage
-    std::lock_guard<std::mutex> lock(cpu_stats_mutex_);
-
-    FILETIME idle_time, kernel_time, user_time;
-    if (!GetSystemTimes(&idle_time, &kernel_time, &user_time)) {
-        return -1.0;
-    }
-
-    // Convert FILETIME to uint64_t (100-nanosecond intervals)
-    auto filetime_to_uint64 = [](const FILETIME& ft) -> uint64_t {
-        return (static_cast<uint64_t>(ft.dwHighDateTime) << 32) | ft.dwLowDateTime;
-    };
-
-    uint64_t idle = filetime_to_uint64(idle_time);
-    uint64_t kernel = filetime_to_uint64(kernel_time); // Includes idle time
-    uint64_t user = filetime_to_uint64(user_time);
-
-    // Kernel time includes idle time, so subtract it to get actual kernel time
-    uint64_t total = kernel + user;
-    uint64_t total_idle = idle;
-
-    if (last_cpu_stats_.total > 0) {
-        uint64_t idle_diff = total_idle - last_cpu_stats_.total_idle;
-        uint64_t total_diff = total - last_cpu_stats_.total;
-
-        last_cpu_stats_.total_idle = total_idle;
-        last_cpu_stats_.total = total;
-
-        if (total_diff > 0) {
-            return ((total_diff - idle_diff) * 100.0) / total_diff;
-        }
-    }
-
-    last_cpu_stats_.total_idle = total_idle;
-    last_cpu_stats_.total = total;
-    return 0.0; // First call, no delta yet
-
-#elif defined(__APPLE__)
-    // macOS: Could use host_processor_info or top command
-    return -1.0; // Not implemented yet
-
+#if defined(__linux__) || defined(_WIN32)
+    return metrics_platform_->get_cpu_usage(cpu_stats_mutex_,
+                                            last_cpu_stats_.total,
+                                            last_cpu_stats_.total_idle);
 #else
-    return -1.0;
+    uint64_t dummy_total = 0, dummy_idle = 0;
+    std::mutex dummy_mutex;
+    return metrics_platform_->get_cpu_usage(dummy_mutex, dummy_total, dummy_idle);
 #endif
 }
 
 // Get GPU usage percentage (AMD GPUs on Linux)
 double Server::get_gpu_usage() {
-#ifdef __linux__
-    // Linux: Read from AMD sysfs (gpu_busy_percent)
-    // Check all GPUs and return the highest utilization
-    try {
-        std::string drm_path = "/sys/class/drm";
-
-        if (!fs::exists(drm_path)) {
-            return -1.0;
-        }
-
-        double highest_usage = -1.0;
-
-        for (const auto& entry : fs::directory_iterator(drm_path)) {
-            std::string card_name = entry.path().filename().string();
-            if (card_name.find("card") != 0 || card_name.find("-") != std::string::npos) {
-                continue;
-            }
-
-            std::string busy_path = entry.path().string() + "/device/gpu_busy_percent";
-            std::ifstream busy_file(busy_path);
-            if (busy_file.is_open()) {
-                double usage;
-                busy_file >> usage;
-                busy_file.close();
-                if (usage > highest_usage) {
-                    highest_usage = usage;
-                }
-            }
-        }
-
-        return highest_usage;
-    } catch (...) {
-        return -1.0;
-    }
-
-#else
-    // GPU usage monitoring not implemented for Windows/macOS
-    return -1.0;
-#endif
+    return metrics_platform_->get_gpu_usage();
 }
 
 // Get VRAM/GTT usage in GB (AMD GPUs on Linux)
 double Server::get_vram_usage() {
-#ifdef __linux__
-    // Linux: Read from AMD sysfs
-    // For dGPU: return VRAM used
-    // For APU: return VRAM + GTT used
-    // On multi-GPU systems, return memory from GPU with highest utilization
-    try {
-        std::string drm_path = "/sys/class/drm";
-
-        if (!fs::exists(drm_path)) {
-            return -1.0;
-        }
-
-        double highest_usage = -1.0;
-        std::string highest_card;
-        double highest_card_memory = 0.0;
-
-        for (const auto& entry : fs::directory_iterator(drm_path)) {
-            std::string card_name = entry.path().filename().string();
-            if (card_name.find("card") != 0 || card_name.find("-") != std::string::npos) {
-                continue;
-            }
-
-            std::string device_path = entry.path().string() + "/device";
-
-            // Read GPU utilization to find the most active GPU
-            double gpu_usage = 0.0;
-            std::ifstream busy_file(device_path + "/gpu_busy_percent");
-            if (busy_file.is_open()) {
-                busy_file >> gpu_usage;
-                busy_file.close();
-            }
-
-            // Check if this is a dGPU (has board_info) or APU (no board_info)
-            bool is_dgpu = fs::exists(device_path + "/board_info");
-
-            // Read VRAM used
-            uint64_t vram_used = 0;
-            std::ifstream vram_file(device_path + "/mem_info_vram_used");
-            if (vram_file.is_open()) {
-                vram_file >> vram_used;
-                vram_file.close();
-            }
-
-            // Read GTT used
-            uint64_t gtt_used = 0;
-            std::ifstream gtt_file(device_path + "/mem_info_gtt_used");
-            if (gtt_file.is_open()) {
-                gtt_file >> gtt_used;
-                gtt_file.close();
-            }
-
-            // Skip if no memory info found
-            if (vram_used == 0 && gtt_used == 0) {
-                continue;
-            }
-
-            // Calculate memory for this card
-            uint64_t card_memory = is_dgpu ? vram_used : (vram_used + gtt_used);
-
-            // Track the GPU with highest utilization
-            if (gpu_usage > highest_usage || highest_usage < 0) {
-                highest_usage = gpu_usage;
-                highest_card = card_name;
-                highest_card_memory = card_memory / (1024.0 * 1024.0 * 1024.0); // Convert to GB
-            }
-        }
-
-        return highest_card_memory > 0 ? highest_card_memory : -1.0;
-    } catch (...) {
-        return -1.0;
-    }
-
-#else
-    // VRAM monitoring not implemented for Windows/macOS
-    return -1.0;
-#endif
+    return metrics_platform_->get_vram_usage_gb();
 }
 
 // Helper: Get NPU utilization (AMD NPU on Linux)
 double Server::get_npu_utilization() {
-#ifdef __linux__
-    try {
-        std::string accel_path = "/dev/accel/accel0";
-        if (!fs::exists(accel_path)) {
-            return -1.0;
-        }
-
-        int fd = open(accel_path.c_str(), O_RDWR);
-        if (fd < 0) {
-            return -1.0;
-        }
-
-        // Check DRM API version (must be 0.7 or later for these IOCTLs)
-        struct drm_version drm_v;
-        memset(&drm_v, 0, sizeof(drm_v));
-        bool version_ok = false;
-        if (ioctl(fd, DRM_IOCTL_VERSION, &drm_v) == 0) {
-            if (drm_v.version_major > 0 || (drm_v.version_major == 0 && drm_v.version_minor >= 7)) {
-                version_ok = true;
-            }
-        }
-
-        if (!version_ok) {
-            close(fd);
-            return -1.0;
-        }
-
-        // Check power_state to avoid waking the NPU if it is asleep
-        fs::path power_state_path = "/sys/class/accel/accel0/device/power_state";
-        if (fs::exists(power_state_path)) {
-            std::ifstream power_file(power_state_path);
-            std::string state;
-            if (power_file >> state) {
-                if (state != "D0") {
-                    close(fd);
-                    return 0.0;
-                }
-            }
-        }
-
-        amdxdna_drm_query_sensor sensors[16] = {};
-        amdxdna_drm_get_info get_info = {};
-        get_info.param = DRM_AMDXDNA_QUERY_SENSORS;
-        get_info.buffer_size = sizeof(sensors);
-        get_info.buffer = (uintptr_t)sensors;
-
-        if (ioctl(fd, DRM_IOCTL_AMDXDNA_GET_INFO, &get_info) < 0) {
-            close(fd);
-            return -1.0;
-        }
-
-        close(fd);
-
-        int num_sensors = get_info.buffer_size / sizeof(amdxdna_drm_query_sensor);
-        double usage_sum = 0.0;
-        int usage_count = 0;
-        for (int i = 0; i < num_sensors; ++i) {
-            if (sensors[i].type == AMDXDNA_SENSOR_TYPE_COLUMN_UTILIZATION) {
-                double val = (double)sensors[i].input * std::pow(10.0, sensors[i].unitm);
-                usage_sum += val;
-                usage_count++;
-            }
-        }
-
-        if (usage_count > 0) {
-            // Return average utilization percentage [0, 100]
-            return (usage_sum / usage_count);
-        }
-
-        return -1.0;
-    } catch (...) {
-        return -1.0;
-    }
-#else
-    // NPU monitoring not implemented for Windows/macOS
-    return -1.0;
-#endif
+    return metrics_platform_->get_npu_utilization();
 }
 
 void Server::handle_system_stats(const httplib::Request& req, httplib::Response& res) {
@@ -4110,49 +3846,7 @@ void Server::handle_system_stats(const httplib::Request& req, httplib::Response&
     stats["cpu_percent"] = (cpu_percent >= 0) ? nlohmann::json(cpu_percent) : nlohmann::json();
 
     // Get memory info
-#ifdef _WIN32
-    MEMORYSTATUSEX memInfo;
-    memInfo.dwLength = sizeof(MEMORYSTATUSEX);
-    if (GlobalMemoryStatusEx(&memInfo)) {
-        double used_gb = (memInfo.ullTotalPhys - memInfo.ullAvailPhys) / (1024.0 * 1024.0 * 1024.0);
-        stats["memory_gb"] = std::round(used_gb * 10.0) / 10.0;
-    } else {
-        stats["memory_gb"] = 0;
-    }
-#elif defined(__linux__)
-    // Linux: Read /proc/meminfo
-    std::ifstream meminfo("/proc/meminfo");
-    if (meminfo.is_open()) {
-        std::string line;
-        long long total_kb = 0, available_kb = 0;
-        while (std::getline(meminfo, line)) {
-            if (line.find("MemTotal:") == 0) {
-                sscanf(line.c_str(), "MemTotal: %lld kB", &total_kb);
-            } else if (line.find("MemAvailable:") == 0) {
-                sscanf(line.c_str(), "MemAvailable: %lld kB", &available_kb);
-                break;
-            }
-        }
-        meminfo.close();
-        double used_gb = (total_kb - available_kb) / (1024.0 * 1024.0);
-        stats["memory_gb"] = std::round(used_gb * 10.0) / 10.0;
-    } else {
-        stats["memory_gb"] = 0;
-    }
-#elif defined(__APPLE__)
-    // macOS: Get memory info
-    int64_t physical_memory = 0;
-    size_t length = sizeof(physical_memory);
-    if (sysctlbyname("hw.memsize", &physical_memory, &length, nullptr, 0) == 0) {
-        // For now, just report total memory since getting free memory is complex on macOS
-        double total_gb = physical_memory / (1024.0 * 1024.0 * 1024.0);
-        stats["memory_gb"] = std::round(total_gb * 10.0) / 10.0;
-    } else {
-        stats["memory_gb"] = 0;
-    }
-#else
-    stats["memory_gb"] = 0;
-#endif
+    stats["memory_gb"] = metrics_platform_->get_memory_usage_gb();
 
     // GPU usage
     double gpu_percent = get_gpu_usage();
