+++ src/cpp/server/utils/path_utils.cpp
@@ -1,4 +1,5 @@
 #include <lemon/utils/path_utils.h>
+#include <lemon/utils/path_platform.h>
 #include <lemon/utils/json_utils.h>
 #include <lemon/utils/process_manager.h>
 #include <algorithm>
@@ -12,13 +13,6 @@
 #include <windows.h>
 #else
 #include <unistd.h>
-#include <limits.h>
-#ifdef __APPLE__
-#include <unistd.h>
-#include <sys/types.h>
-#include <pwd.h>
-#include <mach-o/dyld.h>
-#endif
 #endif
 
 namespace fs = std::filesystem;
@@ -33,6 +27,12 @@ namespace lemon::utils {
 static std::string g_cache_dir;
 static std::string g_models_dir;
 
+// Platform abstraction instance (created on first use)
+static PathPlatform* platform() {
+    static std::unique_ptr<PathPlatform> p = create_path_platform();
+    return p.get();
+}
+
 void set_cache_dir(const std::string& dir) {
     g_cache_dir = dir;
 }
@@ -41,93 +41,20 @@ void set_models_dir(const std::string& dir) {
     g_models_dir = dir;
 }
 
-#ifdef _WIN32
-static std::wstring utf8_to_wstring(const std::string& str) {
-    if (str.empty()) return std::wstring();
-
-    int size_needed = MultiByteToWideChar(CP_UTF8, 0, str.c_str(), -1, nullptr, 0);
-    if (size_needed <= 0) {
-        return std::wstring();
-    }
-
-    std::wstring result(size_needed, L'\0');
-    MultiByteToWideChar(CP_UTF8, 0, str.c_str(), -1, &result[0], size_needed);
-    result.resize(size_needed - 1);
-    return result;
-}
-
-static std::string wstring_to_utf8(const std::wstring& wstr) {
-    if (wstr.empty()) return std::string();
-
-    int size_needed = WideCharToMultiByte(CP_UTF8, 0, wstr.c_str(), -1, nullptr, 0, nullptr, nullptr);
-    if (size_needed <= 0) {
-        return std::string();
-    }
-
-    std::string result(size_needed, '\0');
-    WideCharToMultiByte(CP_UTF8, 0, wstr.c_str(), -1, &result[0], size_needed, nullptr, nullptr);
-    result.resize(size_needed - 1);
-    return result;
-}
-#endif
-
 std::string get_environment_variable_utf8(const std::string& name) {
-#ifdef _WIN32
-    std::wstring wide_name = utf8_to_wstring(name);
-    DWORD size_needed = GetEnvironmentVariableW(wide_name.c_str(), nullptr, 0);
-    if (size_needed == 0) {
-        return "";
-    }
-
-    std::wstring value(size_needed, L'\0');
-    GetEnvironmentVariableW(wide_name.c_str(), &value[0], size_needed);
-    value.resize(size_needed - 1);
-    return wstring_to_utf8(value);
-#else
-    const char* value = std::getenv(name.c_str());
-    return value ? std::string(value) : "";
-#endif
+    return platform()->get_environment_variable_utf8(name);
 }
 
 fs::path path_from_utf8(const std::string& path) {
-#ifdef _WIN32
-    return fs::u8path(path);
-#else
-    return fs::path(path);
-#endif
+    return platform()->path_from_utf8(path);
 }
 
 std::string path_to_utf8(const fs::path& path) {
-#ifdef _WIN32
-    return wstring_to_utf8(path.wstring());
-#else
-    return path.string();
-#endif
+    return platform()->path_to_utf8(path);
 }
 
 std::string get_executable_dir() {
-#ifdef _WIN32
-    char buffer[MAX_PATH];
-    GetModuleFileNameA(NULL, buffer, MAX_PATH);
-    fs::path exe_path(buffer);
-    return exe_path.parent_path().string();
-#elif defined(__linux__)
-    char buffer[PATH_MAX];
-    ssize_t len = readlink("/proc/self/exe", buffer, sizeof(buffer) - 1);
-    if (len != -1) {
-        buffer[len] = '\0';
-        fs::path exe_path(buffer);
-        return exe_path.parent_path().string();
-    }
-#elif defined(__APPLE__)
-    char buffer[PATH_MAX];
-    uint32_t size = sizeof(buffer);
-    if (_NSGetExecutablePath(buffer, &size) == 0) {
-        fs::path exe_path(buffer);
-        return exe_path.parent_path().string();
-    }
-#endif
-    throw std::runtime_error("Unable to resolve executable directory");
+    return platform()->get_executable_dir();
 }
 
 std::string get_resource_path(const std::string& relative_path) {
@@ -139,30 +66,14 @@ std::string get_resource_path(const std::string& relative_path) {
         return resource_path.string();
     }
 
-#ifndef _WIN32
-    // On Linux/macOS, also check standard install locations
-    std::vector<std::string> install_prefixes = {
-        "/Library/Application Support/Lemonade",  // macOS system install location
-        "/usr/local/share/lemonade-server",
-        "/opt/share/lemonade-server",
-        "/usr/share/lemonade-server"
-    };
-
-
-    // Also check user's local install directory
-    const char* home = std::getenv("HOME");
-    if (home) {
-        std::string user_local = std::string(home) + "/.local/share/lemonade-server";
-        install_prefixes.insert(install_prefixes.begin(), user_local);
-    }
-
+    // Check platform-specific install locations
+    std::vector<std::string> install_prefixes = platform()->get_install_prefixes();
     for (const auto& prefix : install_prefixes) {
         fs::path installed_path = fs::path(prefix) / relative_path;
         if (fs::exists(installed_path)) {
             return installed_path.string();
         }
     }
-#endif
 
     // Fallback: return original path (will fail but with clear error)
     return resource_path.string();
@@ -327,65 +238,11 @@ std::string get_cache_dir() {
     }
 
     // Fallback to platform-specific defaults (for backward compat / CLI client)
-#ifdef _WIN32
-    std::string userprofile = get_environment_variable_utf8("USERPROFILE");
-    if (!userprofile.empty()) {
-        return userprofile + "\\.cache\\lemonade";
-    }
-    throw std::runtime_error("USERPROFILE is not set; cannot resolve Lemonade cache directory");
-#elif defined(__APPLE__)
-    if (geteuid() != 0) {
-        std::string home = get_environment_variable_utf8("HOME");
-        if (!home.empty()) {
-            std::string cache_dir = home + "/.cache/lemonade";
-            fs::path cache_path = path_from_utf8(cache_dir);
-            if (!fs::exists(cache_path)) {
-                fs::create_directories(cache_path);
-            }
-            return cache_dir;
-        }
-        struct passwd* pw = getpwuid(getuid());
-        if (pw) {
-            std::string cache_dir = std::string(pw->pw_dir) + "/.cache/lemonade";
-            fs::path cache_path = path_from_utf8(cache_dir);
-            if (!fs::exists(cache_path)) {
-                fs::create_directories(cache_path);
-            }
-            return cache_dir;
-        }
-    }
-
-    {
-        std::string cache_dir = "/Library/Application Support/lemonade/.cache";
-        fs::path cache_path = path_from_utf8(cache_dir);
-        if (!fs::exists(cache_path)) {
-            fs::create_directories(cache_path);
-        }
-        return cache_dir;
-    }
-#else
-    std::string home = get_environment_variable_utf8("HOME");
-    if (!home.empty()) {
-        return home + "/.cache/lemonade";
-    }
-    throw std::runtime_error("HOME is not set; cannot resolve Lemonade cache directory");
-#endif
+    return platform()->get_cache_dir(g_cache_dir);
 }
 
 std::string default_hf_cache_dir() {
-#ifdef _WIN32
-    std::string userprofile = get_environment_variable_utf8("USERPROFILE");
-    if (!userprofile.empty()) {
-        return userprofile + "\\.cache\\huggingface\\hub";
-    }
-    throw std::runtime_error("USERPROFILE is not set; cannot resolve HuggingFace cache directory");
-#else
-    std::string home = get_environment_variable_utf8("HOME");
-    if (!home.empty()) {
-        return home + "/.cache/huggingface/hub";
-    }
-    throw std::runtime_error("HOME is not set; cannot resolve HuggingFace cache directory");
-#endif
+    return platform()->default_hf_cache_dir();
 }
 
 std::string resolve_hf_cache_dir() {
@@ -420,50 +277,7 @@ std::string get_hf_cache_dir() {
 }
 
 std::string get_runtime_dir() {
-#ifdef _WIN32
-    char temp_path[MAX_PATH];
-    GetTempPathA(MAX_PATH, temp_path);
-    return std::string(temp_path);
-#elif defined(__APPLE__)
-    std::error_code ec;
-    fs::path base = fs::temp_directory_path(ec);
-    if (!ec && !base.empty()) {
-        fs::path lemon_dir = base / "lemonade";
-        ec.clear();
-        fs::create_directory(lemon_dir, ec);
-        if (!ec || fs::is_directory(lemon_dir)) {
-            return lemon_dir.string();
-        }
-    }
-    throw std::runtime_error("Unable to resolve writable runtime directory on macOS");
-#else
-    const char* xdg = std::getenv("XDG_RUNTIME_DIR");
-    if (xdg && xdg[0] != '\0') {
-        std::error_code ec;
-        fs::path base(xdg);
-        if (fs::is_directory(base, ec) && !ec && access(xdg, W_OK) == 0) {
-            fs::path lemon_dir = base / "lemonade";
-            ec.clear();
-            fs::create_directory(lemon_dir, ec);
-            std::error_code ec2;
-            if (!ec || fs::is_directory(lemon_dir, ec2)) {
-                return lemon_dir.string();
-            }
-        }
-    }
-
-    // System services can get a RuntimeDirectory= without XDG_RUNTIME_DIR.
-    if (const char* runtime_dir = std::getenv("RUNTIME_DIRECTORY");
-        runtime_dir && runtime_dir[0] != '\0') {
-        std::error_code ec;
-        fs::path base(runtime_dir);
-        if (fs::is_directory(base, ec) && !ec && access(runtime_dir, W_OK | X_OK) == 0) {
-            return base.string();
-        }
-    }
-
-    throw std::runtime_error("Unable to resolve writable runtime directory from XDG_RUNTIME_DIR or RUNTIME_DIRECTORY");
-#endif
+    return platform()->get_runtime_dir();
 }
 
 std::string get_downloaded_bin_dir() {
